@@ -1,27 +1,45 @@
 {
-    "config": {
-        "vendor-dir": "vendor"
+    "name": "lcs/project",
+    "type": "project",
+    "description": ".",
+    "keywords": [""],
+    "license": "MIT",
+    "require": {
+        "php": "^8.0",
+        "league/oauth2-google": "^4.0",
+        "league/plates": "^3.4",
+        "nikic/fast-route": "^1.3",
+        "pecee/simple-router": "4.3.7.2",
+        "phpmailer/phpmailer": "^6.5",
+        "vlucas/phpdotenv": "^5.4"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9"
     },
     "autoload": {
         "psr-4": {
             "App\\": "app/",
-            "Database\\": "database",
-            "Tests\\" : "tests/",
-            "Migrations\\"  : "migrations/"
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/",
+            "Bootstrap\\Database\\": "bootstrap/database/"
         },
         "files": [
             "app/Helpers/Helpers.php"
         ]
     },
-    "require": {
-        "php": "^8.0",
-        "pecee/simple-router": "4.3.7.2",
-        "phpmailer/phpmailer": "^6.5",
-        "vlucas/phpdotenv": "^5.4",
-        "league/oauth2-google": "^4.0",
-        "jenssegers/blade": "^1.4"
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/"
+        }
     },
-    "require-dev": {
-        "phpunit/phpunit": "^9"
-    }
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true,
+        "platform": {
+            "php": "8.1"
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
