@@ -10,7 +10,7 @@ class Session
     public function __construct()
     {
         if (!session_id()) {
-           // session_start();
+           //session_start();
         };
     }
 
