@@ -1,5 +1,9 @@
 <?php
 
 return [
-  'url' => env('APP_URL', 'http://localhost')
+  'url' => env('APP_URL', 'http://localhost'),
+
+  'timezone' => 'America/Sao_Paulo',
+
+  'locale' => 'pt-BR',
 ];
