@@ -1,5 +1,5 @@
 const mix = require('laravel-mix');
 
-mix.copyDirectory('resources/assets/css','public/css');
-mix.copyDirectory('resources/assets/js','public/js');
-mix.copyDirectory('resources/assets/images','public/images');
+mix.copyDirectory('resources/assets/css','public/assets/css');
+mix.copyDirectory('resources/assets/js','public/assets/js');
+mix.copyDirectory('resources/assets/images','public/assets/images');
